Move remaining `Gen[*Entity]` from `Generators` to `GenDBModel`
