+++ app/src/test/scala/org/alephium/explorer/Generators.scala
@@ -30,11 +30,10 @@ import org.alephium.explorer.GenApiModel._
 import org.alephium.explorer.GenCoreApi._
 import org.alephium.explorer.GenCoreProtocol._
 import org.alephium.explorer.GenCoreUtil._
-import org.alephium.explorer.GenDBModel._
 import org.alephium.explorer.api.model._
 import org.alephium.explorer.persistence.model._
 import org.alephium.protocol.{model => protocol, _}
-import org.alephium.protocol.model.{Address, BlockHash, TransactionId}
+import org.alephium.protocol.model.{BlockHash, TransactionId}
 import org.alephium.serde._
 import org.alephium.util.{AVector, Duration, TimeStamp}
 
@@ -42,101 +41,6 @@ import org.alephium.util.{AVector, Duration, TimeStamp}
 object Generators {
 
   def groupSettingGen: Gen[GroupSetting] = Gen.choose(2, 4).map(groupNum => GroupSetting(groupNum))
-
-  @SuppressWarnings(Array("org.wartremover.warts.DefaultArguments"))
-  def transactionEntityGen(blockHash: Gen[BlockHash] = blockHashGen): Gen[TransactionEntity] =
-    for {
-      hash              <- transactionHashGen
-      blockHash         <- blockHash
-      timestamp         <- timestampGen
-      chainFrom         <- groupIndexGen
-      chainTo           <- groupIndexGen
-      gasAmount         <- Gen.posNum[Int]
-      gasPrice          <- u256Gen
-      order             <- Gen.posNum[Int]
-      mainChain         <- Arbitrary.arbitrary[Boolean]
-      scriptExecutionOk <- Arbitrary.arbitrary[Boolean]
-      coinbase          <- Arbitrary.arbitrary[Boolean]
-    } yield
-      TransactionEntity(
-        hash              = hash,
-        blockHash         = blockHash,
-        timestamp         = timestamp,
-        chainFrom         = chainFrom,
-        chainTo           = chainTo,
-        gasAmount         = gasAmount,
-        gasPrice          = gasPrice,
-        order             = order,
-        mainChain         = mainChain,
-        scriptExecutionOk = scriptExecutionOk,
-        inputSignatures   = None,
-        scriptSignatures  = None,
-        coinbase          = coinbase
-      )
-
-  def mempoolTransactionEntityGen(): Gen[MempoolTransactionEntity] =
-    for {
-      hash      <- transactionHashGen
-      chainFrom <- groupIndexGen
-      chainTo   <- groupIndexGen
-      gasAmount <- Gen.posNum[Int]
-      gasPrice  <- u256Gen
-      timestamp <- timestampGen
-    } yield
-      MempoolTransactionEntity(
-        hash      = hash,
-        chainFrom = chainFrom,
-        chainTo   = chainTo,
-        gasAmount = gasAmount,
-        gasPrice  = gasPrice,
-        lastSeen  = timestamp
-      )
-
-  val blockHeaderGen: Gen[BlockHeader] =
-    blockHeaderGenWithDefaults()
-
-  @SuppressWarnings(Array("org.wartremover.warts.DefaultArguments"))
-  def blockHeaderGenWithDefaults(chainFrom: Gen[GroupIndex] = groupIndexGen,
-                                 chainTo: Gen[GroupIndex]   = groupIndexGen,
-                                 height: Gen[Height]        = heightGen): Gen[BlockHeader] =
-    for {
-      hash         <- blockHashGen
-      timestamp    <- timestampGen
-      chainFrom    <- chainFrom
-      chainTo      <- chainTo
-      height       <- height
-      version      <- Gen.posNum[Byte]
-      depStateHash <- hashGen
-      txsHash      <- hashGen
-      txsCount     <- Gen.posNum[Int]
-      nonce        <- bytesGen
-      target       <- bytesGen
-      hashrate     <- arbitrary[Long].map(BigInteger.valueOf)
-      mainChain    <- Arbitrary.arbitrary[Boolean]
-      parent       <- Gen.option(blockHashGen)
-    } yield
-      BlockHeader(
-        hash         = hash,
-        timestamp    = timestamp,
-        chainFrom    = GroupIndex.unsafe(chainFrom.value),
-        chainTo      = GroupIndex.unsafe(chainTo.value),
-        height       = Height.unsafe(height.value),
-        mainChain    = mainChain,
-        nonce        = nonce,
-        version      = version,
-        depStateHash = depStateHash,
-        txsHash      = txsHash,
-        txsCount     = txsCount,
-        target       = target,
-        hashrate     = hashrate,
-        parent       = parent
-      )
-
-  val blockHeaderTransactionEntityGen: Gen[(BlockHeader, List[TransactionEntity])] =
-    for {
-      blockHeader <- blockHeaderGen
-      transaction <- Gen.listOf(transactionEntityGen(Gen.const(blockHeader.hash)))
-    } yield (blockHeader, transaction)
 
   def parentIndex(chainTo: GroupIndex)(implicit groupSetting: GroupSetting) =
     groupSetting.groupNum - 1 + chainTo.value
@@ -320,213 +224,6 @@ object Generators {
       )
     })
   }
-
-  def blockHeaderWithHashrate(timestamp: TimeStamp, hashrate: Double): Gen[BlockHeader] = {
-    for {
-      hash         <- blockHashGen
-      from         <- groupIndexGen
-      to           <- groupIndexGen
-      height       <- heightGen
-      nonce        <- bytesGen
-      version      <- Gen.posNum[Byte]
-      depStateHash <- hashGen
-      txsHash      <- hashGen
-      txsCount     <- Gen.posNum[Int]
-      target       <- bytesGen
-      parent       <- Gen.option(blockHashGen)
-    } yield {
-      BlockHeader(
-        hash,
-        timestamp,
-        from,
-        to,
-        height,
-        true,
-        nonce,
-        version,
-        depStateHash,
-        txsHash,
-        txsCount,
-        target,
-        BigDecimal(hashrate).toBigInt.bigInteger,
-        parent
-      )
-    }
-  }
-
-  /** Update toUpdate's primary key to be the same as `original` */
-  def copyPrimaryKeys(original: BlockDepEntity, toUpdate: BlockDepEntity): BlockDepEntity =
-    toUpdate.copy(
-      hash = original.hash,
-      dep  = original.dep
-    )
-
-  val blockDepGen: Gen[BlockDepEntity] =
-    for {
-      hash  <- blockHashGen
-      dep   <- blockHashGen
-      order <- Gen.posNum[Int]
-    } yield
-      BlockDepEntity(
-        hash  = hash,
-        dep   = dep,
-        order = order
-      )
-
-  /** Generates a tuple2 of [[BlockDepEntity]] where the second one has
-    * the same primary key as the first one but with different values */
-  val blockDepUpdatedGen: Gen[(BlockDepEntity, BlockDepEntity)] =
-    for {
-      dep1 <- blockDepGen
-      dep2 <- blockDepGen
-    } yield {
-      (dep1, copyPrimaryKeys(dep1, dep2))
-    }
-
-  /**
-    * Table `uinputs` applies uniqueness on `(output_ref_key, tx_hash)`.
-    *
-    * Table `uoutputs` applies uniqueness on `(tx_hash, address, uoutput_order)`.
-    *
-    * These tables allow `txHashes` and `addresses` to have duplicate data
-    * and some of our SQL queries run `DISTINCT` on these columns.
-    * This function generates test data for such queries.
-    *
-    * @return Generator that result in data that allows multiple [[UInputEntity]]
-    *         with the same/duplicate `txHashes` and/or `addresses`.
-    */
-  def duplicateTxIdAndAddressGen(): Gen[(Gen[TransactionId], Gen[Address])] =
-    for {
-      txHashes <- Gen.nonEmptyListOf(transactionHashGen)
-      address  <- Gen.nonEmptyListOf(addressGen)
-    } yield (Gen.oneOf(txHashes), Gen.oneOf(address))
-
-  def uinputEntityWithDuplicateTxIdsAndAddressesGen(): Gen[List[UInputEntity]] =
-    for {
-      (txHash, addressOpt) <- duplicateTxIdAndAddressGen()
-      entities             <- Gen.listOf(uinputEntityGen(txHash, Gen.option(addressOpt)))
-    } yield entities
-
-  def uoutputEntityWithDuplicateTxIdsAndAddressesGen(): Gen[List[UOutputEntity]] =
-    for {
-      (txHash, addressOpt) <- duplicateTxIdAndAddressGen()
-      entities             <- Gen.listOf(uoutputEntityGen(txHash, addressOpt))
-    } yield {
-      //uoutput allows duplicate `tx_hash` and `address` and there are queries
-      //that search for such data but these duplicates must have unique `uoutputOrder`.
-      //This sets all uoutputs to have have unique `uoutputOrder`.
-      entities.zipWithIndex map {
-        case (entity, index) =>
-          entity.copy(uoutputOrder = index)
-      }
-    }
-
-  def uInOutEntityWithDuplicateTxIdsAndAddressesGen()
-    : Gen[(List[UInputEntity], List[UOutputEntity])] =
-    for {
-      (txHash, addressOpt) <- duplicateTxIdAndAddressGen()
-      inputs               <- Gen.listOf(uinputEntityGen(txHash, Gen.option(addressOpt)))
-      outputs              <- Gen.listOf(uoutputEntityGen(txHash, addressOpt))
-    } yield
-      (inputs, outputs.zipWithIndex map {
-        case (entity, index) =>
-          entity.copy(uoutputOrder = index)
-      })
-
-  /** Update toUpdate's primary key to be the same as `original` */
-  def copyPrimaryKeys(original: InputEntity, toUpdate: InputEntity): InputEntity =
-    toUpdate.copy(
-      outputRefKey = original.outputRefKey,
-      txHash       = original.txHash,
-      blockHash    = original.blockHash
-    )
-
-  /** Generate two InputEntities where the second one is an update of the first */
-  @SuppressWarnings(Array("org.wartremover.warts.DefaultArguments"))
-  def updatedInputEntityGen(
-      outputEntityGen: Gen[OutputEntity] = outputEntityGen): Gen[(InputEntity, InputEntity)] =
-    for {
-      input1 <- inputEntityGen(outputEntityGen)
-      input2 <- inputEntityGen(outputEntityGen)
-    } yield {
-      (input1, copyPrimaryKeys(input1, input2))
-    }
-
-  /** Update toUpdate's primary key to be the same as `original` */
-  def copyPrimaryKeys(original: OutputEntity, toUpdate: OutputEntity): OutputEntity =
-    toUpdate.copy(
-      key       = original.key,
-      blockHash = original.blockHash
-    )
-
-  /** Generates a tuple2 of [[OutputEntity]] where the second one has
-    * the same primary key as the first one but with different values */
-  def updatedOutputEntityGen(): Gen[(OutputEntity, OutputEntity)] =
-    for {
-      output1 <- outputEntityGen
-      output2 <- outputEntityGen
-    } yield {
-      (output1, copyPrimaryKeys(output1, output2))
-    }
-
-  /** Update toUpdate's primary key to be the same as `original` */
-  def copyPrimaryKeys(original: TransactionEntity, toUpdate: TransactionEntity): TransactionEntity =
-    toUpdate.copy(hash = original.hash, blockHash = original.blockHash)
-
-  /** Generates a [[BlockEntity]] with optional parent */
-  @SuppressWarnings(Array("org.wartremover.warts.DefaultArguments"))
-  def updatedTransactionEntityGen(
-      blockHash: Gen[BlockHash] = blockHashGen): Gen[(TransactionEntity, TransactionEntity)] =
-    for {
-      transaction1 <- transactionEntityGen(blockHash)
-      transaction2 <- transactionEntityGen(blockHash)
-    } yield {
-      (transaction1, copyPrimaryKeys(transaction1, transaction2))
-    }
-
-  /** Update toUpdate's primary key to be the same as original */
-  def copyPrimaryKeys(original: BlockHeader, toUpdate: BlockHeader): BlockHeader =
-    toUpdate.copy(
-      hash         = original.hash,
-      depStateHash = original.depStateHash
-    )
-
-  /** Generates a tuple2 of [[BlockHeader]] where the second one has
-    * the same primary key as the first one but with different values */
-  def updatedBlockHeaderGen(): Gen[(BlockHeader, BlockHeader)] =
-    for {
-      blockHeader1 <- blockHeaderGen
-      blockHeader2 <- blockHeaderGen
-    } yield {
-      (blockHeader1, copyPrimaryKeys(blockHeader1, blockHeader2))
-    }
-
-  /**
-    * Generates a [[BlockEntity]] with optional parent
-    * @param randomMainChainGen Some if randomness to mainChain should be applies else None.
-    *                           The generated boolean mainChain is set for both parent and child [[BlockEntity]].
-    */
-  @SuppressWarnings(Array("org.wartremover.warts.DefaultArguments"))
-  def genBlockEntityWithOptionalParent(
-      groupIndexGen: Gen[GroupIndex]           = Gen.const(GroupIndex.unsafe(0)),
-      randomMainChainGen: Option[Gen[Boolean]] = None): Gen[(BlockEntity, Option[BlockEntity])] =
-    for {
-      groupIndex <- groupIndexGen
-      parent     <- Gen.option(blockEntityWithParentGen(groupIndex, groupIndex, None))
-      child      <- blockEntityWithParentGen(groupIndex, groupIndex, parent)
-    } yield {
-      randomMainChainGen.flatMap(_.sample) match {
-        case Some(randomMainChain) =>
-          //set main_chain to generated sample
-          val childWithRandomMainChain  = child.copy(mainChain = randomMainChain)
-          val parentWithRandomMainChain = parent.map(_.copy(mainChain = randomMainChain))
-
-          (childWithRandomMainChain, parentWithRandomMainChain)
-
-        case None =>
-          (child, parent)
-      }
-    }
 
   /**
     * Generates [[Pagination]] instance for the generated data.
