+++ app/src/test/scala/org/alephium/explorer/persistence/queries/UInputQueriesSpec.scala
@@ -21,7 +21,7 @@ import slick.jdbc.PostgresProfile.api._
 
 import org.alephium.explorer.{AlephiumFutureSpec, GenCommon}
 import org.alephium.explorer.GenCoreProtocol.transactionHashGen
-import org.alephium.explorer.Generators._
+import org.alephium.explorer.GenDBModel._
 import org.alephium.explorer.persistence.{DatabaseFixtureForAll, DBRunner}
 import org.alephium.explorer.persistence.model._
 import org.alephium.explorer.persistence.schema._
