+++ app/src/test/scala/org/alephium/explorer/service/TransactionHistoryServiceSpec.scala
@@ -24,7 +24,7 @@ import slick.jdbc.PostgresProfile.api._
 
 import org.alephium.explorer.AlephiumFutureSpec
 import org.alephium.explorer.ConfigDefaults._
-import org.alephium.explorer.Generators._
+import org.alephium.explorer.GenDBModel._
 import org.alephium.explorer.api.model._
 import org.alephium.explorer.persistence.{DatabaseFixtureForEach, DBRunner}
 import org.alephium.explorer.persistence.model.TransactionEntity
